+++ src/test/java/study/proxy/cglib/HelloCglibProxyTest.java
@@ -1,22 +1,20 @@
-package study.proxy;
+package study.proxy.cglib;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.lang.reflect.Proxy;
+import org.springframework.cglib.proxy.Enhancer;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-public class HelloJdkDynamicProxyTest {
+public class HelloCglibProxyTest {
     @Test
-    @DisplayName("Jdk Dynamic Proxy 학습 테스트")
-    void jdkDynamicProxyTest() {
-        Hello helloProxy = (Hello) Proxy.newProxyInstance(
-                Hello.class.getClassLoader(),
-                new Class[]{Hello.class},
-                new HelloUpperCaseInvocationHandler(new HelloTarget())
-        );
+    @DisplayName("Cglib Proxy 학습 테스트")
+    void cglibProxyTest() {
+        Enhancer enhancer = new Enhancer();
+        enhancer.setSuperclass(HelloTarget.class);
+        enhancer.setCallback(new HelloUpperCaseMethodInterceptor(new HelloTarget()));
+        HelloTarget helloProxy = (HelloTarget) enhancer.create();
 
         assertAll(
                 () -> assertThat(helloProxy.sayHello("wu2ee")).isEqualTo("HELLO WU2EE"),
