+++ src/test/java/study/proxy/cglib/HelloTarget.java
@@ -1,17 +1,14 @@
-package study.proxy;
+package study.proxy.cglib;
 
-public class HelloTarget implements Hello {
-    @Override
+public class HelloTarget {
     public String sayHello(String name) {
         return "Hello " + name;
     }
 
-    @Override
     public String sayHi(String name) {
         return "Hi " + name;
     }
 
-    @Override
     public String sayThankYou(String name) {
         return "Thank You " + name;
     }
