+++ src/test/java/study/proxy/jdkdynamic/Hello.java
@@ -1,4 +1,4 @@
-package study.proxy;
+package study.proxy.jdkdynamic;
 
 public interface Hello {
     String sayHello(String name);
