+++ src/test/java/study/proxy/jdkdynamic/HelloJdkDynamicProxyTest.java
@@ -1,20 +1,22 @@
-package study.cglib;
+package study.proxy.jdkdynamic;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.cglib.proxy.Enhancer;
+
+import java.lang.reflect.Proxy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-public class HelloCglibProxyTest {
+public class HelloJdkDynamicProxyTest {
     @Test
-    @DisplayName("Cglib Proxy 학습 테스트")
-    void cglibProxyTest() {
-        Enhancer enhancer = new Enhancer();
-        enhancer.setSuperclass(HelloTarget.class);
-        enhancer.setCallback(new HelloUpperCaseMethodInterceptor(new HelloTarget()));
-        HelloTarget helloProxy = (HelloTarget) enhancer.create();
+    @DisplayName("Jdk Dynamic Proxy 학습 테스트")
+    void jdkDynamicProxyTest() {
+        Hello helloProxy = (Hello) Proxy.newProxyInstance(
+                Hello.class.getClassLoader(),
+                new Class[]{Hello.class},
+                new HelloUpperCaseInvocationHandler(new HelloTarget())
+        );
 
         assertAll(
                 () -> assertThat(helloProxy.sayHello("wu2ee")).isEqualTo("HELLO WU2EE"),
