+++ src/test/java/study/proxy/jdkdynamic/HelloTarget.java
@@ -1,14 +1,17 @@
-package study.cglib;
+package study.proxy.jdkdynamic;
 
-public class HelloTarget {
+public class HelloTarget implements Hello {
+    @Override
     public String sayHello(String name) {
         return "Hello " + name;
     }
 
+    @Override
     public String sayHi(String name) {
         return "Hi " + name;
     }
 
+    @Override
     public String sayThankYou(String name) {
         return "Thank You " + name;
     }
